+++ AutoSuggestBox/AutoSuggestBoxSample/AutoSuggestBoxSamples/CustomRendering.xaml
@@ -1,0 +1,13 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:local="clr-namespace:AutoSuggestBoxSample.AutoSuggestBoxSamples"
+             x:Class="AutoSuggestBoxSample.AutoSuggestBoxSamples.CustomRendering">
+    <ContentPage.Content>
+        <StackLayout>
+            <Label Text="This uses a custom renderer and override the border style on iOS" HorizontalOptions="Start" />
+            <local:CustomAutoSuggestBox x:Name="SuggestBox1" 
+                           PlaceholderText="This has a full border on iOS" />
+        </StackLayout>
+    </ContentPage.Content>
+</ContentPage>
